server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
redis
pydantic
python-dotenv
pytest
